imx: imx8mm: kontron: Add support for PMIC PCA9450
In order to configure the PMIC at an early stage, we add the driver
for SPL. The registers to set are taken from the NXP EVK in
downstream uboot-imx. We only do the configuration that might be
essential for booting the kernel. We will let the kernel driver
handle all the advanced settings for voltage scaling in suspend mode,
etc.

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>