imx8mm-kontron-bl: Fix operation with USH-I card
The imx8mm-energy-manager board fails boot U-Boot after a reset.

From Kontron's SL_i.MX8M_Mini_User_Guide_EN.pdf:

"If UHS-I mode is required uSDHC2 should be used as this interface is
powered from a dedicated LDO and the supply can be switched from +3.3 V
to +1.8 V. This voltage NVCC_SD is accessible on pin B8 to
connect external pull-ups."

On the imx8mm-energy-manager board the NVCC_SD is left unconnected
and the board is not being able to properly switch to 1.8V to achieve
UHS-I speeds.

To workaround the problem, remove the vqmmc-supply property, so that
NVCC_SD does not switch to 1.8V.

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Luan Rafael Carneiro <luan.rafael@ossystems.com.br>