sl-mx8mm: Assign the eth1 MAC address
Currently, the eth1 MAC address is not assigned, which causes the kernel
to generate a random one for the Ethernet USB port.

Fix this problem by assigning the eth1 MAC address as the eth0
MAC address + 1.

The original implementation made by Kontron did not work
as it returned the MAC address of 00:00:00:00:00:00.

Signed-off-by: Fabio Estevam <festevam@gmail.com>
Signed-off-by: Luan Rafael Carneiro <luan.rafael@ossystems.com.br>